\name{permFDP-package}
\alias{permFDP-package}
\alias{permFDP}
\docType{package}
\title{
  Rejection threshold correction using permutation-based FDP estimation.
}
\description{
  <Paragraph under construction.>
}
\details{
  <Section under construction.>
}
\author{
Steven R. Shuken

Maintainer: Steven R. Shuken <srshuken@gmail.com>
}
\references{
  Shuken and McNerney 2021
}
\keyword{ package }
\seealso{
  <Under construction>
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
