% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permFDP.R
\name{permFDP.adjust.threshold}
\alias{permFDP.adjust.threshold}
\title{Permutation-Based FDP Method for Rejection Threshold Correction}
\usage{
permFDP.adjust.threshold(pVals, threshold, myDesign, intOnly, nPerms)
}
\arguments{
\item{pVals}{Vector of p-values. The length of this vector must be the same as the number of rows in the intOnly data frame}

\item{threshold}{Original threshold that will be adjusted.}

\item{myDesign}{Vector of 1s and 2s specifying which columns in intOnly belong to the control group (1) and which belong to the test group (2).}

\item{intOnly}{A data frame. Each column is a sample, each row is an analyte.}

\item{nPerms}{The number of permutations to perform. At least 100 is recommended.}
}
\description{
This function controls FDR using the permutation method described in our manuscript. Like the BH method above, it corrects the rejection threshold rather than the p-values themselves.
It returns the new threshold for P-value rejection.
It uses the Rcpp and BH packages to leverage fast C++ code.
}
\examples{
cat_function()
}
\keyword{FDP,}
\keyword{FDR,}
\keyword{p-values,}
\keyword{permutation}
